requirements.txt
praw
python-dotenv
openai
moviepy
requests
